refactor(plugin): keep X-Forwarded-For helper next to ServeHTTP
