+++ pkg/modsecurity/serve.go
@@ -3,6 +3,7 @@ package modsecurity
 import (
 	"bytes"
 	"io"
+	"net"
 	"net/http"
 	"strconv"
 	"strings"
@@ -215,4 +216,19 @@ func forwardResponse(resp *http.Response, rw http.ResponseWriter) {
 	if _, err := io.Copy(rw, resp.Body); err != nil {
 		return
 	}
+}
+
+// appendPeerToXForwardedFor appends the host part of remoteAddr to header's X-Forwarded-For.
+// remoteAddr is the incoming peer as host:port. Parse failure leaves the header unchanged.
+func appendPeerToXForwardedFor(header http.Header, remoteAddr string) {
+	clientIP, _, err := net.SplitHostPort(remoteAddr)
+	if err != nil {
+		return
+	}
+	// Join any existing chain, then append this hop — same as ReverseProxy Director.
+	prior := header.Values("X-Forwarded-For")
+	if len(prior) > 0 {
+		clientIP = strings.Join(prior, ", ") + ", " + clientIP
+	}
+	header.Set("X-Forwarded-For", clientIP)
 }
